2006-03-12  Alexis Sukrieh  <sukria@backup-manager.org>
	* lib/files.sh:
	  + Applyied patch 21 for closing #50. 
	    (Stephen Kitt)




git-svn-id: svn://svn.backup-manager.org/backup-manager/trunk@368 2e458433-d701-0410-9826-f9b593394a3c

+++ lib/files.sh
@@ -193,12 +193,14 @@ clean_file()
 	fi
 
 	date=$(get_date_from_file $file)
-	date=$(echo $date | sed -e 's/[^0-9]//g')
-	if [ ! -z $date ]; then
-		if [ $date -lt $date_to_remove ] || 
-		   [ $date = $date_to_remove ]; then
-			info "Removing \$file"
-			rm -f $file
+	if [ $date != $file ]; then
+		date=$(echo $date | sed -e 's/[^0-9]//g')
+		if [ ! -z $date ]; then
+			if [ $date -lt $date_to_remove ] || 
+			   [ $date = $date_to_remove ]; then
+				info "Removing \$file"
+				rm -f $file
+			fi
 		fi
 	fi
 }
